// src/services/fop/fop.interface.ts
export interface IFop {
  id: number;
  first_fop_text: string;
  second_fop_text: string;
  third_fop_text: string;
  fourth_fop_text: string;
  first_date_fop: string;
  second_date_fop: string;
  code_edr_fop: string;
  register_date_fop: string;
  bank_fop: string;
  number_fop: string;
  address_fop: string;
}